# openapi.json
{"openapi":"3.1.0","info":{"title":"MatchAI API","version":"0.1.0"},"paths":{"/register":{"post":{"summary":"Registrar Usuario","operationId":"registrar_usuario_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UsuarioCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UsuarioResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/login":{"post":{"summary":"Login","operationId":"login_login_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_login_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/job-offers/job-offers":{"get":{"tags":["Jobs"],"summary":"Listar Ofertas","operationId":"listar_ofertas_job_offers_job_offers_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/documents/upload":{"post":{"tags":["documents"],"summary":"Upload Document","operationId":"upload_document_api_documents_upload_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_document_api_documents_upload_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/documents/":{"get":{"tags":["documents"],"summary":"Obtener Documentos","operationId":"obtener_documentos_api_documents__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/career/analyze":{"post":{"tags":["career"],"summary":"Analyze","operationId":"analyze_api_career_analyze_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CareerRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/career/rank-jobs":{"get":{"tags":["career"],"summary":"Rank Jobs","operationId":"rank_jobs_api_career_rank_jobs_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/career/optimized-cv/{job_offer_id}":{"get":{"tags":["career"],"summary":"Obtener Cv Optimizado","operationId":"obtener_cv_optimizado_api_career_optimized_cv__job_offer_id__get","parameters":[{"name":"job_offer_id","in":"path","required":true,"schema":{"type":"integer","title":"Job Offer Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/career/cover-letter/{job_offer_id}":{"get":{"tags":["career"],"summary":"Obtener Cover Letter","operationId":"obtener_cover_letter_api_career_cover_letter__job_offer_id__get","parameters":[{"name":"job_offer_id","in":"path","required":true,"schema":{"type":"integer","title":"Job Offer Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/career/application-answers/{job_offer_id}":{"get":{"tags":["career"],"summary":"Obtener Application Answers","operationId":"obtener_application_answers_api_career_application_answers__job_offer_id__get","parameters":[{"name":"job_offer_id","in":"path","required":true,"schema":{"type":"integer","title":"Job Offer Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/career/interview-preparation/{job_offer_id}":{"get":{"tags":["career"],"summary":"Obtener Interview Preparation","operationId":"obtener_interview_preparation_api_career_interview_preparation__job_offer_id__get","parameters":[{"name":"job_offer_id","in":"path","required":true,"schema":{"type":"integer","title":"Job Offer Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/career/dashboard":{"get":{"tags":["career"],"summary":"Dashboard","operationId":"dashboard_api_career_dashboard_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/career/job-preferences":{"get":{"tags":["career"],"summary":"Get Job Preferences","operationId":"get_job_preferences_api_career_job_preferences_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserJobPreferenceResponse"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"put":{"tags":["career"],"summary":"Update Job Preferences","operationId":"update_job_preferences_api_career_job_preferences_put","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserJobPreferenceCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserJobPreferenceResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/career/search-jobs":{"get":{"tags":["career"],"summary":"Search Jobs","operationId":"search_jobs_api_career_search_jobs_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/interview/start":{"post":{"tags":["Interview"],"summary":"Start","operationId":"start_api_interview_start_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"role","in":"query","required":true,"schema":{"type":"string","title":"Role"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/interview/answer":{"post":{"tags":["Interview"],"summary":"Answer","operationId":"answer_api_interview_answer_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"session_id","in":"query","required":true,"schema":{"type":"integer","title":"Session Id"}},{"name":"answer","in":"query","required":true,"schema":{"type":"string","title":"Answer"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Home","operationId":"home__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Body_login_login_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"^password$"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","format":"password","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"format":"password","title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_login_post"},"Body_upload_document_api_documents_upload_post":{"properties":{"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_document_api_documents_upload_post"},"CareerRequest":{"properties":{"job_offer_id":{"type":"integer","title":"Job Offer Id"}},"type":"object","required":["job_offer_id"],"title":"CareerRequest"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"UserJobPreferenceCreate":{"properties":{"desired_role":{"type":"string","title":"Desired Role"},"target_countries":{"items":{"type":"string"},"type":"array","title":"Target Countries"},"target_cities":{"items":{"type":"string"},"type":"array","title":"Target Cities"},"remote":{"type":"boolean","title":"Remote","default":true},"hybrid":{"type":"boolean","title":"Hybrid","default":false},"onsite":{"type":"boolean","title":"Onsite","default":false},"published_within_days":{"type":"integer","title":"Published Within Days","default":7}},"type":"object","required":["desired_role","target_countries","target_cities"],"title":"UserJobPreferenceCreate"},"UserJobPreferenceResponse":{"properties":{"desired_role":{"type":"string","title":"Desired Role"},"target_countries":{"items":{"type":"string"},"type":"array","title":"Target Countries"},"target_cities":{"items":{"type":"string"},"type":"array","title":"Target Cities"},"remote":{"type":"boolean","title":"Remote","default":true},"hybrid":{"type":"boolean","title":"Hybrid","default":false},"onsite":{"type":"boolean","title":"Onsite","default":false},"published_within_days":{"type":"integer","title":"Published Within Days","default":7},"id":{"type":"integer","title":"Id"},"user_id":{"type":"integer","title":"User Id"}},"type":"object","required":["desired_role","target_countries","target_cities","id","user_id"],"title":"UserJobPreferenceResponse"},"UsuarioCreate":{"properties":{"email":{"type":"string","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"UsuarioCreate"},"UsuarioResponse":{"properties":{"id":{"type":"integer","title":"Id"},"email":{"type":"string","title":"Email"},"creado_en":{"type":"string","format":"date-time","title":"Creado En"}},"type":"object","required":["id","email","creado_en"],"title":"UsuarioResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"login"}}}}}}